refactored test arrange code
