+++ app/Testing/Concerns/MakesHttpRequestsEx.php
@@ -3,6 +3,11 @@ namespace App\Testing\Concerns;
 
 trait MakesHttpRequestsEx
 {
+    protected function getBaseUri(): string
+    {
+        return "/";
+    }
+
     /**
      * Visit the given URI with a JSON GET request.
      * @param string $uri
@@ -12,5 +17,17 @@ trait MakesHttpRequestsEx
     public function getJson(string $uri, array $headers = []): self
     {
         return $this->json('GET', $uri, [], $headers);
+    }
+
+    /**
+     * @param array $params
+     * @return array
+     */
+    public function getJsonResponse(array $params): array
+    {
+        $parameters = http_build_query($params);
+        $uri = $this->getBaseUri() . "?" . $parameters;
+        $this->getJson($uri);
+        return $this->response->json();
     }
 }
