+++ database/factories/JikanModelFactory.php
@@ -13,11 +13,16 @@ abstract class JikanModelFactory extends Factory
      */
     public function definition(): array
     {
+        return $this->serializeStateDefinition($this->definitionInternal());
+    }
+
+    protected function serializeStateDefinition($stateDefinition): array
+    {
         /**
          * @var Serializer $serializer
          */
         $serializer = app("SerializerV4");
-        return $serializer->toArray($this->definitionInternal());
+        return $serializer->toArray($stateDefinition);
     }
 
     protected abstract function definitionInternal(): array;
